requirements.txt
faiss-cpu
sentence-transformers
pandas
numpy
statsmodels
fastapi
uvicorn
streamlit
